Update patch set 4
Patch Set 4:

I discussed this per email and in person with Aleksey. I would love to have it like this:

 BuildItem::withId()
     ->andLabel( 'foo' )
     ->andSiteLink( 'enwiki', 'foo' )
     ->build();

I like this quite a lot. Main advantages: The "Build…" naming scheme allows for nice code completion. The "and…" naming scheme makes it very obvious that these can't be the first method in the chain.

Instead of "build" the last call can also be "create" or something like this.

Instead of "with…" it can also be "having…",

No global functions please as long as we can not manage functions in namespaces. MediaWiki core does have a very strict "mw…()" naming scheme for global functions. An "item()" function would be way to generic. It might not cause actual trouble, but I find it odd and a bit scary. When we have namespaces for functions we can talk about this again.

Patch-set: 4
Label: Verified=0