Update patch set 1
Patch Set 1: Code-Review+1

The new behavior seems fine, and should fix the issue at hand. Needs manual verification, though.

A side note: skipping the -{...}- syntax if the values are the same for all languages makes sense, but really has nothing to do with the URL syntax. The fact that this change fixes the syntax problem is kind of a happy accident: Plain URL values are the same for all target languages. If that assumption breaks, the syntax problem will be back.

A more robust way to avoid the URL parsing issue would be to follow the URL with <nowiki></nowiki>.

Patch-set: 1