Refactor EntityHandler to get rid of getContentClass
This is a prerequisit for properly doing Ie62bd06.

This patch should be straight refactoring that does not change
anything. What this patch essentially does is:
* removing the bad "getContentClass" that returns a class name.
* turning "makeEntityContent" into an abstract method subclasses must
  implement.
* change "supportsRedirects" to return false by default. Subclasses
  must override it, instead of relying on the feature detection if a
  "newFromRedirect" method exists.
* change "makeEntityRedirectContent" to return null by default.
  Subclasses that support redirects must override it.

Note this must also be changed in the Lexeme as well as MediaInfo
extension. Patches are already uploaded and should receive a +2 before
this here is merged.

Change-Id: Id4fe528d122e05738a8f5d80842d4521587f8d3c