Update patch set 1
Patch Set 1:

I tried to find a proper solution without adding extra burden and it seems even in core it doesn't check for user being blocked in page view and it only does it when the user tries to edit. I'm not sure about keeping this functionality as it buys us some database performance and in it's not used in core too (see first part of my comment) but not so sure about removing it either.

Patch-set: 1
Reviewer: Ladsgroup <920@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0