Update patch set 6
Patch Set 6: Code-Review-1

(1 comment)

I looked at the code again, and was confused by the way the change is tested. See my longer comment on OtherProjectsSidebarGeneratorTest.

The actual change to production code seems OK as far as I managed to test it.
I would like to see some more explanation how this is tested, or test case(s) to be changed somehow if you think this is needed.

Patch-set: 6
Label: Code-Review=-1