Update patch set 15
Patch Set 15:

page id might be an option, although this will mean SqlEntityIdPager would need to be changed to make it possible to set the start offset.

I guess for this script it is critical to be able to re-start from the a given point as it might not be possible/intended to rebuild it all at one go. Or am I wrong with that?

Patch-set: 15
Label: Verified=0