Update patch set 2
Patch Set 2:

Just a remark: The gray icons usually mean the feature is disabled. I believe it was a bug that one of the two icons appeared to be disabled when clicked. But this patch fixed the inconsistency the other way around, by making both icons appear disabled when clicked. I found this surprising.

However, I can see the benefit of this solution. This gives additional feedback when the icons are clicked and the popups open. This is basically a 4th state (normal, disabled, hover, and active), it just uses the same color as the disabled state. Quickly talked with PM. Lydia and I agree it's a good solution as it is now. I just wanted to document this here.

Patch-set: 2