Better names for WikibaseRepo/Client (de)serializer factory methods
This is a resubmission of Ic126c46.

The problem: We do have a "Internal Serialization" and a "DataModel
Serialization" component. The names are confusing.

The so called "DataModel Serialization" component is a "base"
(de)serialization component that only supports Items and Properties.
This is on purpose. We should not teach this component how to deal
with other entity types.

Warning, the renamed method is used in WikibaseQualityConstraints!
Please +2 the patch in that extension first.

Bug: T160439
Change-Id: Idd4f43abec3e6b9930b0cc17138243ca1b9668b6