Update patch set 1
Patch Set 1:

I had an other look at this change and was wondering: Is there anything I can do to make this smaller and easier to review? For example, add a test that makes it obvious what is happening in this patch? But this is pure refactoring in the sense that no test is affected. The essence of this patch is a removed protected method (which can not be tested separately), and it makes a method in an abstract base class abstract (again, this was not possible to test before, and is not possible to test after the change).

The only "controversial" change is the fact that I removed some @see tags I found unnecessary.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0