Make TermSqlIndexBuilder only rebuild entity terms when necessary
When all terms of an entity are there, there is not duplicates,
and the full entity ID column is populated (when present in the DB),
TermSqlIndexBuilder will not remove and add the same set of terms.

This should avoid redundant write DB operations. Writes will
only happen for entities that really need their term data rebuilt.
On the flip side, if there is a need for write operation,
term data will be read twice (once in TermSqlIndexBuilder, and
then in TermSqlIndex::saveTermsOfEntity).

Bug: T162533
Change-Id: I59e0f2db5a3ba2f968647ed6522b26c181e2e88a