Only load entity data TermSqlIndexBuilder when necessary
Makes TermSqlIndexBuilder only load entity data and check
if any of terms are missing or are outdated when requested.

This should make the "defult" rebuild faster, as it
will only check the contents of wb_terms to detect possible
unpopulated ID fields and/or duplicate term entries,
without needing to load all the data for each
processed entity.

RebuildTermSqlIndex maintenance script also gets the
respective option, which is turned off by default.

Bug: T162533
Change-Id: Ia0409ebfd7c905b178c8ebe17945226a845e010c