Update patch set 6
Patch Set 6:

(5 comments)

I've had a brief look at the production code (i.e. haven't really looked at tests yet).
The only question I'd had is why hook handling is kind of happening ChangesListLinesHandler and not in ClientHooks (see inline comment, might be also something that Ladsgroup was referring to). Is it in order to make testing easier, or even possible at all? To be clear, I am in general not really sure if putting this into ClientHooks would be ultimately better just because it would be in ClientHooks. I am simply interested in the motivation why that way.

In general it looks like it is going in the right direction! I am going to test it locally soon. You might be able to elaborate on ChangesListLinesHandler vs ClientHooks in the meanwhile

Patch-set: 6