Update patch set 1
Patch Set 1:

I disagree with the premise of this change. When I see $content->getItem(), it'S immdiately clear that this is guaranteed to return an Item. When I see $content->getEntity(), this is not clear at all. And even if it's $itemContent->getEntity(), I'm still bound to go and check what guarantees getEntity actually gives.

I vote for keeping the specific getters.

I want CR-X "I don't like it" :)

Patch-set: 1
Label: Code-Review=0