Update patch set 1
Patch Set 1:

I don't think having those methods causes any harm, so I have no problem with keeping them. OTOH there seems also to be no huge win because of the fact they're here.

>  I'm still bound to go and check what guarantees getEntity actually gives.

That's true. If we want to have a method that will guarantee (by guarantee I mean a real guarantee, not a documentation) to its callers that it returns Item, not any Entity.
Not sure if this is the case looking at the use of methods (i.e. is there really a need for such method), so I am more on the position "Whatever".

> I want CR-X "I don't like it" :)

I believe it is called CR-2 "Do not submit" :)

Patch-set: 1
Reviewer: WMDE-leszek <2626@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0