Update patch set 6
Patch Set 6:

(1 comment)

> The class' structure follows a pattern I learned during development of EchoNotificationsHandlers: static method instantiating the class and passing all arguments to a public (testable) method with similar name (on.../do.../test...).

Yeah, that seems like a reasonable pattern!
I was just confused why most of the hooks are just all packed into ClientHooks.
I've talked to Daniel and you explained it to me that that all things in ClientHooks are simply old code, and the approach you use here is the preferred one. So good job!

Thanks Matěj for explanations! I will try to test this patch locally in the next hours.

Patch-set: 6
Label: Verified=0