Update patch set 2
Patch Set 2:

I find "media" problematic. There is an actual "Media:" keyword in MediaWiki. This is a pseudo namespace that can be used to link to file description pages instead of getting a thumbnail.

The special page we are using here is called [[Special:FilePath]]. Why not use something like this? Name the constant MEDIA_FILE_BASE_URI and the function getMediaFileURI()?

Patch-set: 2
Label: Code-Review=0