Add hooks for saving and removing statements
The wikibase.statement.saved hook fires whenever a statement (new or
edited) is saved, and the wikibase.statement.removed hook fires whenever
a statement is removed, corresponding directly to the two methods of
StatementsChanger. Both hooks receive two arguments: the entity ID and
the statement GUID of the affected statement.

Please note that these hooks are not necessarily stable: like the rest
of the Wikibase JavaScript code, the are an Unstable Interface [1] and
subject to change without notice.

[1]: https://www.wikidata.org/wiki/Wikidata:Stable_Interface_Policy#Unstable_Interfaces

Bug: T167870
Change-Id: I5efdb67c57106c64dd2bee0e888ba1bb230886ea