Update patch set 1
Patch Set 1: Code-Review+1

I like the idea of having hooks for better integration of widgets.
The problem I see here is that it introduces a mediaWiki dependency which might not be acceptable. We might find a better place for the hook or at least make it a 'soft dependency'.

Patch-set: 1
Label: Code-Review=+1