Update patch set 2
Patch Set 2: Code-Review+1

I almost wanted to merge PS1. It works as expected, the code is nice. But I was wondering: How can this code rely on the local this.options.value being updated? Do we guarantee this is always the case *before* stopEditing is called? And: Why is the CSS class set *between* two stopEditing calls for the sub-widgets a statement view is made of? It should happen at the end, shouldn't it? But there is this return $.when( … ), which makes it hard to have stuff after.

Please have a look at PS2. This is my suggestion how this might look like. Feel free to change whatever you want, I will review it again. Or just click +2 if it's ok for you. :-)

Patch-set: 2
Label: Code-Review=+1