Update patch set 3
Patch Set 3:

The reason for the test failures is simple: the two failing tests instantiate StatementsChanger with incomplete arguments, so the entity is undefined…

I’m not sure what the right fix is. Should I update the tests to supply some fake entity, or handle the possibility of a missing entity in StatementsChanger?

Patch-set: 3