Update patch set 3
Patch Set 3:

> The reason for the test failures is simple: the two failing tests
 > instantiate StatementsChanger with incomplete arguments, so the
 > entity is undefined…
 > 
 > I’m not sure what the right fix is. Should I update the tests to
 > supply some fake entity, or handle the possibility of a missing
 > entity in StatementsChanger?

As soon as EntityChanger clearly states `@param {wikibase.datamodel.Entity} entity` in constructor, it means that it doesn't accept `null` or whatever.
So the actual problem is that those tests are violating constructor interface, so getting to the point:
Yes, tests should be fixed - it is their problem.

Patch-set: 3
Label: Verified=0