Update patch set 4
Patch Set 4:

I updated the tests. The object still isn’t a real Entity, but at least it’s an object. Is there an easy way to get a mock / fake Entity? (I looked around for a bit, but the EntityTermsChanger tests in the same directory also use object literals with just the necessary members instead of proper Entity objects.)

Patch-set: 4
Label: Verified=0