Update patch set 4
Patch Set 4: Code-Review+1

(2 comments)

I'm not happy with the soft dependency in a place where nobody expects it, and nobody will find it ever again. I'm willing to merge this as it is for now because it does the job, and my concern is only about the internal structure of the code, not about the fact that these exact hooks are fired at these exact places.

A proper solution is to inject a hook handling function as a constructor parameter. That function would look something like this:

 var fireHook = mediaWiki && typeof mediaWiki.hook === 'function'
     ? function ( name ) {
           mediaWiki.hook( name ).fire.apply( this, arguments.slice( 1 ) );
       }
     : function () {};

The benefit of this is that this soft dependency is in a single, central place, and not scattered around the code base. The downside is that this must be injected multiple levels deep.

As said, maybe something for a later patch.

Patch-set: 4