Update patch set 5
Patch Set 5: Code-Review+1

(1 comment)

I updated this to use the class name "NewItem" and follow the naming scheme "…::withLabel( … )->andLabel( … )". This means all methods are duplicated, as a static one named "with…" and a non-static one named "and…".

I can already see quite a lot of possible expansions, like the one Lucas mentioned. Or support for more value types that are commonly used in tests. Badges are also missing. A "NewStatement" builder class is missing that supports qualifiers and references. And so on. All for later patches.

This is fine for a +2 in my opinion.

Patch-set: 5
Label: Code-Review=+1