Update patch set 2
Patch Set 2: Code-Review+1

The code looks trivial and good, but I wonder. What is the actual issue this patch aims to fix? Is there a Phabricator ticket that describes the issue? If not, can you create one that helps us understand? Or, instead of a ticket, is it possible to add a test that shows what happens if this code is called on a namespace where Wikibase access is *not* enabled? Having such a test is quite critical (if it's possible to create one).

Patch-set: 2