Update patch set 74
Patch Set 74:

Testing on http://elastic-wikidata.wmflabs.org shows that there is a minor issue: When an alias is matched, it is correctly shown in addition to the preferred label. However, if a label in a fallback language is matched (e.g. a search in de-ch matches the en label), that matched label is not shown in addition to the preferred (perhaps de) label.

I'm not 100% sure the faulty code is in this patch, but it should be investigated.

Patch-set: 74