Only apply “disabled” style to inner spans
novalue snaks and other kinds of snaks have the wikibase-snakview-…
class on both the surrounding wikibase-snakview-value div and the inner
span. Only apply the “disabled” style (gray, italic) to the inner span,
so that any elements added after it within the div (e. g. the constraint
report gadget) do not get this style applied.

Change-Id: I8f2783d4bf01d0fb083f79d3e29e0e5ed86d0164