Update patch set 2
Patch Set 2:

> The commit message does not explain *why* this move needs to be done. 

Added a line on this.

> This makes this setting inaccessible for clients, right? Is this intended?

I can't see it being used in client currently. And also I don't know in which case it would be useful for a client. Seems to be a repo-specific thing to me. I am not sure about that though, therefore I'd be happy to be told this is wrong interpretation of this setting.

Patch-set: 2
Label: Verified=0