Use the …::class feature instead of "ClassName::staticMethod" strings
The only (but crucial) benefit this gives us is that IDEs can understand
this actually references the class, and we can find usages of these
classes much more easily.

I tried to use the even more expressive array syntax whenever possible.
IDEs like PHPStorm are able to understand this even better than the
string syntax, and can resolve both the class and the method name. But
this is not possible for hooks as well as the Echo stuff. In both cases
the code expects callbacks to be a string, and does something unexpected
when a callback is in array form.

Change-Id: I7c75829b5755b48a6e5e492f9ade8583a6afd4ca