Do not discard RecentChanges lines in case of an error
This was just recently introduced with I62d1b42.

I'm not 100% sure what this patch does is correct. Please review
carefully. My idea is: Whenever one of these hook handlers returns
false, the execution of other hook handlers is stopped and the entire
recent changes line currently processed is not displayed. I doubt this
is what we want. This is hiding an error. This is hiding possibly
crucial information from the user. Even if there is an error, I believe
we can still display the recent changes line as it is, unmodified by
the hooks. I assume it will not look pretty, but displaying something
is still better than displaying noting, I believe.

Change-Id: Ie1e6f4cc0e2066f55bad070a2de97d5418806285