Update patch set 2
Patch Set 2: Code-Review-1

I will merge this, but just realized options.wiki needs an update. I will do this and amend this patch.

I tried to think of a reason to use this specific setting on a client, but can not think of one. Currently it is not used in client code.

Patch-set: 2
Label: Code-Review=-1