Introduce more specific API error messages and allow localization
This patch focuses on a few error messages that got lost in Ib5b1eae.
I'm either introducing new localizable messages for common use cases,
or reuse existing ones.

This also fixes a long standing bug in ChangeOpStatement. This is very
closely related to the other changes because it is also about properly
reporting API errors with localizable messages.

The only minor detail that might be a little bit controversial are the
two places where I reuse the existing "wikibase-api-invalid-property-id"
message. This does not have a parameter. The following patch I0d2a471
will introduce that parameter.

Bug: T168859
Change-Id: I5c26fc968fd800a9df8e65f1cf7e099fd3644648