Populate full entity ID column and only once on wb_terms schema update
Instead of using the multi-purpose term index rebuild script,
just run simpler SQL queries to fill the data the new column.
Also log the update, so the populating of the column is only
run once.

Note: this simpler method can only generate full IDs for items
and properties. It cannot handle any possible custom entity
types.

TODO: Maybe it should also check that wb_terms is not too
much, and skip filling the column if this would take too
long?

TODO: how it should report that the column has not been
populated and the setting should be set to false (as it
is quite critical)?

Bug: T168708
Change-Id: I3be4361217a6a194373ff48f391be30983d512b0