Update patch set 5
Patch Set 5:

I'm afraid I never fully understood what was missing here, because both relevant patches are listed under "Same Topic" here in Gerrit, are both depending on this one, and had both a +2 already.

* I49fb9ef8811782e3972ce9ba0352e798ef52ed02
* I42144baffba28ad749a053f1fe34021bfb9f0641

I rebased all 3 patches now, and added the missing "newEntityContent" method to the 2 sub-patches. Please restore the +2 on these two first, so they are all merged the same time. This is the easiest thing to do. Otherwise a complicated series of at least 6 patches would be needed to always keep the code in a consistent state.

Patch-set: 5