Update patch set 2
Patch Set 2:

PS2: moved running the queries to own class to have code in DatabaseSchemaUpdater nicer.

Note that running the script is still behind the condition onreadFullEntityIdColumn setting.
Here is my reasoning: as the script can in some cases (now only case being there are terms for something else than items and properties in the table, so we cannot build full ids for those; maybe there would be some other cases where the script will not do populating?) can refuse to populate the new column, it is critical to let user now that the column is not populated if readFullEntityIdColumn is not complete -as this would lead to horrible amount of errors and other evilness.

Therefore, currently the script will stop and die in case when it cannot populate column for all term rows. With this, the script will not be logged as run, i.e. it will be run again with another run of update.php
As the script dies with a long message, it might give some help to the owner of the site, what to do, and then possibly re-run update.

If readFullEntityIdColumn is false, update.php should probably finish as usual, even if there are possible issues with contents of wb_terms. In such case some error should maybe be presented to user ("you need to fill the column") which currently does not happen, as the script is not even run if readFullEntityIdColumn is false.

OTOH, I believe this all might be mostly theoretical considerations here. I wouldn't be surprised if only Wikibase instances that have terms for custom entities were our dev boxes. In such case it could possibly be also fine to always run the script, and maybe not even refuse to populate if there are custom entities in wb_terms. Maybe just populate all items and properties, always, no matter what settings are. If there were some stuff that could not be populated, inform user on this (this would be critical as would also like to drop the column during the schema update I think).

What do you think?

Sorry for the long comment.

Patch-set: 2