Make readFullEntityIdColumn enabled by default
The full entity ID column will be populated during schema update
(update.php), so Wikibase instances should be fine to use the
new full ID column. Wikibase instances should in general
use the full ID column unless there is any reason to not
populate the column using update script.

DEPLOYMENT: when deploying this for Wikidata make sure
readFullEntityIdColumn is set to false in local settings,
as Wikidata does not use a script to populate the column.

Change-Id: I8f8c16e7a2e105571ba23d8e157471dc2e668273