Update patch set 1
Patch Set 1:

> Why do you think putting JSON in data attribute is a bad idea?

Not in general. Only when it creates the situation I described. For example, when the user wants the JSON of a specific statement, he can either process the JSON still stored in the wbEntity configuration variable, or call the wbgetclaims API module. We are also talking about a nice ResourceLoader module as a replacement for wbEntity. I believe we should encourage users to either use the API or that ResourceLoader module, and not introduce a fourth method.

Patch-set: 1