Update patch set 1
Patch Set 1:

What do you mean? The PHPDoc blocks touched in this patch add zero information to what is already there from the function header. If the specific documentation generator we are using is unable to generate the same basic information from the function header, this is a problem of the documentation generator. But I doubt this is the case. I can see the function header in the example link you posted.

Patch-set: 1