Update patch set 1
Patch Set 1:

Thiemo suggested it but when I was testing it in localhost I got the warning because my client's language was Persian and the badge in repo didn't have label in Persian. It is a very possible scenario in small wikis and it will flag logs of Wikimedia. I don't think it should be there.

Patch-set: 1
Reviewer: Ladsgroup <920@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0