Update patch set 4
Patch Set 4:

Using "builder" to refer to callback functions that instantiate an object was my mistake. It's misleading, because of the builder pattern.

For MediaWikiServices in core, we decided to call these "instantiator callbacks" or "instantiators". I find that name a lot better. Sorry for the mess...

Patch-set: 4
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0