Remove all default "return true" from repo hooks
When a hook returns true, it means the exact same as returning null.
This is already the case for years. Note that this was already
inconsistent in this file. Some hooks never returned anything, others
had a "return true" in the end, but an "if ( … ) { return; }" with no
bool in their code.

Change-Id: Iaa3c6e7f91c0a6e6a01af54187ceb2240896ac22