Update patch set 2
Patch Set 2:

PS2: Made two changes after discussing this again with Daniel.

1. The populating of the column will only happen when both write- and read full id column settings are set to true.
Rationale behind this: when write setting is set to false, wiki maintainer does not want data in the new column. Then just don't add it during update.php.
Setting the read setting to false in turn is interpreted here as making a statement: "I will populate the new column myself". 

2. I changed the behaviour of the maint script doing the actual populating in case there are custom entity type terms in the table. Now it will print the warning, populate data for items and properties and mark the script as run (so it does not run again).
Not the nicest solution probably but I couldn't think about something better that would not be overthinking the issue. Hopefully the only wikis that have this problem are just a few of our test wikis where there might be mediainfo entity terms. We should be able to handle that.

Patch-set: 2
Label: Verified=0