Rename and rework ApiErrorReporter::dieMessage into dieWithError
This is one of the smallest patches I can think of that is split from
the bigger patch I0d2a471.

This patch reworks the existing (and barely used) method dieMessage into
one that is compatible with ApiBase::dieWithError. However, I'm minifying
the interface as much as possible and only introduce features that are
actually used (e.g. no Message object can be provided, and $errorCode is
not optional and not derived from the message key).

In addition I'm making some PHPDoc comments more precise.

Change-Id: Idb1c008e4360b7a284e1dd83f84f3066e5293d17