Update patch set 4
Patch Set 4: Code-Review-1

(3 comments)

Something I did not reviewed yet are all the messages that are now marked as translatable. Are we sure they are all fine and should be translated? I think these need individual review, and this will take a while.

My -1 is meant as a reminder for myself, because I want to continue working on this patch. It's also meant as an RfC for other devs. What do you think?

Patch-set: 4
Label: Code-Review=-1