Update patch set 3
Patch Set 3:

Hmm, that could work, I am a bit concerned though.
As I see how this class is used, it sometimes gets (in a constructor) a read connection, sometimes a write connection, depending on what its user actually needs (so the parameter name is not accurate, although yeah, the "new" writeConnection is only used in write operations and we kind of "know" it is indeed a write connection when needed).

As I got the rationale from the commit message it is better to open two read connections even if one will not be used, than to use write connection for read operations.

I am not super happy with the approach, but OK if this is what needs to be done fast. But I'd like to have a Tech Debt ticket about cleaning up the way EntityUsageTable is created.

Patch-set: 3