Fix foreign IDs handling in SqlEntityInfoBuilder
Database queries referring to page titles of entity pages,
and to full entity IDs should consider that IDs of entities
consider "foreign" by local repo are not prefixed in the
foreign repo's database.

This also makes SqlEntityInfoBuilder locally rely on numeric entity
IDs only when it is necessary. Otherwise full IDs (without
a prefix when applicable) are used. This makes the class
slightly less bound to the numeric entity IDs only.

SqlEntityInfoBuilderTest::testEntityIdsArePrefixedWithRepositoryNameFullEntityId
should have caught at least part of the issue but the
test setup was done wrong, so test has been passing
misleadingly.

This fixes the test, and adds tests covering the other
part of behaviour related to the issue.

Bug: T169199
Change-Id: I61b8c1a576b600c803284d528f40ae761a117ac6