Update patch set 3
Patch Set 3:

Looking at usage of EntityUsageTable I was wondering about the following:

For example SqlUsageTracker::addUsedEntities works in a way that it reads existing usage data, calculates current usage data, checks what of new is messing in the DB, and puts it back.
All this was intentionally done using single connection, to not miss any data. I wonder (really as in "wonder" - I don't really know and don't know if this is an issue or not) whether it could lead to some problems now, if the "old" data was taken from the possibly not synchronized replica, and data are put to master which could possibly already contain some data or something similar?

Patch-set: 3
Reviewer: WMDE-leszek <2626@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0