Update patch set 3
Patch Set 3:

> would a smaller batch size help when doing the select queries on
 > master?

To be honest, I rather not send any reads to the master unless we specifically do so (ie: giving it some weight on config files). 
If by any means it is possible to stop sending SELECTs to the master, the queries need to be small and fast. 
Ideally the code should be able to deal with small inconsistencies that might happen from having replication.
Even though delays are not happening on a daily basis, it must be assumed that it could happen from time to time, that we have some seconds of delay.
This is the delay history we have had in the last few days:

https://grafana.wikimedia.org/dashboard/db/mysql-aggregated?panelId=6&fullscreen&orgId=1&var-dc=eqiad%20prometheus%2Fops&var-group=core&var-shard=s5&var-role=All&from=now%2Fw&to=now%2Fw

Patch-set: 3