Update patch set 4
Patch Set 4: Code-Review+1

I did a first round of reviewing this patch and love it very much. Thanks a lot for putting so much effort into this! This is now very well tested, much better than ever before. I believe this is really valuable, even if the change that will be visible to users is pretty small.

My comment was just a remark and not meant as an edit request. I would like us (either myself or an other team member, feel free!) to manually test this and merge it then.

Patch-set: 4