Update RecentChange attributes handling for core change Ic3a434c0
Creating a RecentChange object by passing a fake row is somewhat fragile
if the contents of said row ever change, as is happening in Ic3a434c0.

This change updates Wikibase for Ic3a434c0 in a manner that should be
compatible with core both before and after that change.

Change-Id: I4a1236a81098ec0e924c0569a9b146478bbc235c