Update patch set 2
Patch Set 2:

(1 comment)

This is meant as the most simple bugfix, and I don't like this to grow and become more complicated than it needs to be. The message was never seen by anybody anyway. Currently it doesn't surface at all via the API. The unlocalized text is dead code, and all this patch does is removing this dead code. Previously the test was in the API response, but in a custom structure no third party understands. And our own code does not need this particular message.

Also this is a test for the MediaInfo extension, something that is not deployed on any production system anyway. Nobody even tries to use this.

Patch-set: 2