Update patch set 1
Patch Set 1: Code-Review-1

A much more descriptive commit message would be very much welcome. What is the actual problem you want to solve?

Also I believe this patch is incomplete. The method is often called with specific options, and these do not contain the …OBJECTS_FOR_MAPS option. I believe if we want to enable this by default, we must enable it everywhere, not only when the method is called with no options.

Patch-set: 1
Label: Code-Review=-1