Update patch set 1
Patch Set 1:

I would like the JSON serialization returned by the 'revisions' API to be more predictable (and look more like the wbgetentities serialization) as a workaround for T40971.

E.g. you can see that in https://www.wikidata.org/w/api.php?action=query&prop=revisions&revids=516678024&rvprop=content the empty maps are serialized as JSON arrays instead of objects:

    {"type":"item","id":"Q32336167","labels":[],"descriptions":[],"aliases":[],"claims":[],"sitelinks":[]}

The patch seems to do the trick, but of course it is not the proper way...
And I know the 'revisions' serialization is probably not meant to be as stable as the 'wbgetentities' one :(

Patch-set: 1
Label: Verified=0