Remove unused SnakFormatter::FORMAT_HTML_WIDGET
There is a single reference in the Math extension, thats why I keep
the constant as a @deprecated one for a while.

One could argue this is a breaking change because of the removed option
from the wbformatvalue API. I say: Let's not care about this to much.
It's not used anyway. It never ever did anything different that the
HTML mode. But feel free to disagree if you think this needs to go
through the official deprecation process.

Change-Id: Icb446783a6c27090095e66492a75fcb3bb0f66c1