Update patch set 3
Patch Set 3:

Perhaps that actually is an incompatibility with the new version of the dependency? If the test passes with the old version, that seems to be the case. And if this is the case, the questions is - how hard is it to fix this? Or should the change in the library be reverted?

Patch-set: 3
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0