Use HTMLCacheUpdateJob instead of invalidateCache
This change replaces the use of Title::invalidateCache() and
Title::purgeSquid() by HTMLCacheUpdateJob, which does both.
Also, HTMLCacheUpdateJob supports batching, removing the overhead of
scheduling one job per target page (and doing a separate database update
for each target page, too).

The distinction between purging the parser cache and purging the web cache
has been dropped entirely. For the moment, there is never a reason to do one
but not the other: we always want to schedule a page for re-parse via a
RefreshLinksJob if we invalidated the parser cache.

This may change when if we can update cached ParserOutput objects in-place,
without re-parsing. In that case, we would update the ParserCache, instead of
invalidating it. It seems we would have to somehow hook into RefreshLinksJob
to do this properly. Left for later.

DEPLOY: update any dashboard using wikibase.client.pageupdates.* counters
to use the new WebCache.jobs and WebCache.titles counters instead of
the ParserCache and WebCache counters.

Bug: T164173
Change-Id: I3a59486808b699ad0f9463e2af56d69fd37f8bf7