Use HTMLCacheUpdateJob instead of invalidateCache.
This change replaces the use of Title::invalidateCache() and
Title::purgeSquid() by HTMLCacheUpdateJob, which does both.

The distinction between purging the parser cache and purging the web cache
has been dropped entirely. For the moment, there is never a reason to do one
but not the other. This may change when if we can update cached ParserOutput
objects in-place, without re-parsing. In that case, we would update the
ParserCache, instead of invalidating it.

DEPLOY: update any dashboard using wikibase.client.pageupdates.* counters
to use the new WebCache.jobs and WebCache.titles counters.

Bug: T164173
Change-Id: I3a59486808b699ad0f9463e2af56d69fd37f8bf7