Handle storage exception from custom entities in EntityStorageHelper
Also changes the exception thrown by WikiPageEntityStore::assignFreshId
when entity that already hsa an ID is provided. This method expects
an entity without the ID set, so when it is provided something other,
the caller must have been wrong.
All existing callers already make sure the entity passed to that
method actually is the one with the "empty" ID.

Change-Id: Ic78b4db0d24cc230197b7e17691c22ae013a79ba