Update patch set 3
Patch Set 3: Code-Review-1

Hardcoding the page_touched bump seems like a step backwards. If that's what's needed, one should either call Title::invalidateCache() as before, or use HTMLCacheUpdateJob - which is the batch equivalent of performing a CDN purge and calling Title::invalidateCache.

Depending on what triggered a HTMLCacheUpdateJob, a RefreshLinksJob is also enqueued that proactively re-parse the page. Otherwise the re-parse would be deferred until a user views the page. While that works for data shown on the page itself (at the cost of introducing parser in user latency), it doesn't work for secondary data shown on category pages, file pages, WhatLinksHere etc, which is why RefreshLinksJob is prioritised.

Anyhow, I don't know the larger picture of what kind of change is being propagated here, but it seems better to use these jobs instead of touching the database directly which decentralises this logic and creates technical debt. These are quite sensitive database fields that I'd rather not see an extension write to directly.

Patch-set: 3
Label: Code-Review=-1