Update patch set 3
Patch Set 3:

@Krinkle thanks for pointing me to HTMLCacheUpdate, I didn't find it the other day. I agree that we should avoid touching the page table directly.

I don't see when and where the RefreshLinksJobs would be triggered by HTMLCacheUpdate. I suspect that I'll have to do this myself here.

Basically, the changes that are propagated here are changes to wikidata items. When an item changes, all pages using data from that item need to be rerendered. In some cases, rerendering can be avoided by updating the cached ParserOutput object in-place. In that case, we'd still need a CDN purge. 

The current code was written with the idea in mind that we may in some cases want to invalidate the parser cache, but not trigger a RefreshLinksJob. I now believe that was a mistake, we never want that.

Patch-set: 3
Label: Verified=0